Migrate APIRoutesService to TS
